+++ CHANGELOG.md
@@ -11,6 +11,7 @@
 - Migrated `ConfirmRemoveBloodPressureDialog` to Mobius
 - Migrate `UpdatePhoneNumberDialog` to Mobius
 - Separate input fields providers based on country
+- Add separate labels for "Street address" and "Village/Colony/Ward" labels in `PatientEntryScreen` and `EditPatientScreen` for Chennai facility groups
 - [In Progress: 12 Aug 2020] Migrate `TheActivity` to Mobius
 - [In Progress: 11 Aug 2020] Migrate `PatientSearchResultsScreen` to Mobius
 - [In Progress: 12 Aug 2020] Migrate `LinkIdWithPatientView` to Mobius